requirements.txt
albumentations==0.4.6
blobfile==2.0.2
jittor==1.3.7.14
mpi4py==3.1.4
npy_append_array==0.9.16
numpy==1.23.0
omegaconf==2.1.2
opencv_python==4.8.0.74
Pillow==9.5.0
PyYAML==6.0
tqdm==4.65.0